requirements.txt
mysql-connector-python==8.1.0
aioredis==2.0.1
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
asgi-ratelimit==0.10.0
async-timeout==4.0.3
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.103.1
fastapi-limiter==0.1.5
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-resources==6.0.1
iniconfig==2.0.0
limits==3.6.0
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
passlib==1.7.4
pluggy==1.3.0
protobuf==4.21.12
pyasn1==0.5.0
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==1.7.1
pytest==7.4.2
python-jose==3.3.0
python-multipart==0.0.6
redis==4.6.0
requests==2.31.0
rsa==4.9
six==1.16.0
slowapi==0.1.8
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
wrapt==1.15.0
